Move IntrFlags to interrupt.rs
