+++ src/interrupt.rs
@@ -4,6 +4,68 @@ use esp_idf_sys::*;
 
 pub(crate) static CS: IsrCriticalSection = IsrCriticalSection::new();
 
+/// Interrupt allocation flags.
+/// These flags can be used to specify which interrupt qualities the code calling esp_intr_alloc* needs.
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub enum IntrFlags {
+    // Accept a Level 1 interrupt vector (lowest priority)
+    Level1,
+    // Accept a Level 2 interrupt vector.
+    Level2,
+    // Accept a Level 3 interrupt vector.
+    Level3,
+    // Accept a Level 4 interrupt vector.
+    Level4,
+    // Accept a Level 5 interrupt vector.
+    Level5,
+    // Accept a Level 6 interrupt vector.
+    Level6,
+    // Accept a Level 7 interrupt vector (highest priority)
+    Nmi,
+    // Interrupt can be shared between ISRs.
+    Shared,
+    // Edge-triggered interrupt.
+    Edge,
+    // ISR can be called if cache is disabled.
+    // Must be used with a proper option *_ISR_IN_IRAM in SDKCONFIG
+    Iram,
+    // Return with this interrupt disabled.
+    IntrDisabled,
+    // Low and medium prio interrupts. These can be handled in C.
+    LowMed,
+    // High level interrupts. Need to be handled in assembly.
+    High,
+    // Mask for all level flags.
+    LevelMask,
+}
+
+impl From<IntrFlags> for u32 {
+    fn from(flag: IntrFlags) -> Self {
+        match flag {
+            IntrFlags::Level1 => esp_idf_sys::ESP_INTR_FLAG_LEVEL1,
+            IntrFlags::Level2 => esp_idf_sys::ESP_INTR_FLAG_LEVEL2,
+            IntrFlags::Level3 => esp_idf_sys::ESP_INTR_FLAG_LEVEL3,
+            IntrFlags::Level4 => esp_idf_sys::ESP_INTR_FLAG_LEVEL4,
+            IntrFlags::Level5 => esp_idf_sys::ESP_INTR_FLAG_LEVEL5,
+            IntrFlags::Level6 => esp_idf_sys::ESP_INTR_FLAG_LEVEL6,
+            IntrFlags::Nmi => esp_idf_sys::ESP_INTR_FLAG_NMI,
+            IntrFlags::Shared => esp_idf_sys::ESP_INTR_FLAG_SHARED,
+            IntrFlags::Edge => esp_idf_sys::ESP_INTR_FLAG_EDGE,
+            IntrFlags::Iram => esp_idf_sys::ESP_INTR_FLAG_IRAM,
+            IntrFlags::IntrDisabled => esp_idf_sys::ESP_INTR_FLAG_INTRDISABLED,
+            IntrFlags::LowMed => esp_idf_sys::ESP_INTR_FLAG_LOWMED,
+            IntrFlags::High => esp_idf_sys::ESP_INTR_FLAG_HIGH,
+            IntrFlags::LevelMask => esp_idf_sys::ESP_INTR_FLAG_LEVELMASK,
+        }
+    }
+}
+
+impl Default for IntrFlags {
+    fn default() -> Self {
+        Self::Level1
+    }
+}
+
 /// Returns true if the currently active core is executing an ISR request
 #[inline(always)]
 #[link_section = ".iram1.interrupt_active"]
